requirements_core.txt
Django==1.6.5
Pillow==2.5.2
South==1.0

requests==2.3.0
boto==2.32.1
gunicorn==19.1.0
psycopg2==2.5.3
django-storages==1.1.8
dj-database-url==0.3.0

django-bootstrap3==4.10.0
django-oauth2-provider==0.2.6.1
django-registration==1.0
django-braces==1.4.0

# Solves transparency issue
# https://github.com/mariocesar/sorl-thumbnail/pull/274
http://github.com/mariocesar/sorl-thumbnail/archive/master.zip#egg=sorl-thumbnail==11.12.1b
